requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
configparser==4.0.2
docker-pycreds==0.4.0
gitdb==4.0.2
GitPython==3.1.0
gql==0.2.0
graphql-core==1.1
idna==2.9
numpy==1.18.1
nvidia-ml-py3==7.352.0
pathtools==0.1.2
promise==2.3
psutil==5.7.0
python-dateutil==2.8.1
PyYAML==5.3
requests==2.23.0
sentry-sdk==0.14.2
shortuuid==0.5.0
six==1.14.0
smmap==3.0.1
subprocess32==3.5.4
torch==1.4.0
urllib3==1.25.8
wandb==0.8.28
watchdog==0.10.2
